♻️ Rename .../list-example/components to examples
