+++ apps/cookbook/src/app/examples/examples.routes.ts
@@ -22,14 +22,14 @@ import { FormFieldExampleComponent } from './form-field-example/form-field-examp
 import { GridExampleComponent } from './grid-example/grid-example.component';
 import { IconExampleComponent } from './icon-example/icon-example.component';
 import { ItemExampleComponent } from './item-example/item-example.component';
-import { ListBoldTextOnRowSelectionExampleComponent } from './list-example/components/bold-text-on-row-selection';
-import { ListColoredItemsExampleComponent } from './list-example/components/colored-items';
-import { ListWithDividersExampleComponent } from './list-example/components/dividers';
-import { ListWithHeaderAndFooterExampleComponent } from './list-example/components/header-and-footer';
-import { ListItemsExampleComponent } from './list-example/components/items';
-import { ListWithSectionsExampleComponent } from './list-example/components/sections';
-import { ListWithSectionsAndColoredItemsExampleComponent } from './list-example/components/sections-and-colored-items';
-import { ListSelectableItemsExampleComponent } from './list-example/components/selectable-items';
+import { ListBoldTextOnRowSelectionExampleComponent } from './list-example/examples/bold-text-on-row-selection';
+import { ListColoredItemsExampleComponent } from './list-example/examples/colored-items';
+import { ListWithDividersExampleComponent } from './list-example/examples/dividers';
+import { ListWithHeaderAndFooterExampleComponent } from './list-example/examples/header-and-footer';
+import { ListItemsExampleComponent } from './list-example/examples/items';
+import { ListWithSectionsExampleComponent } from './list-example/examples/sections';
+import { ListWithSectionsAndColoredItemsExampleComponent } from './list-example/examples/sections-and-colored-items';
+import { ListSelectableItemsExampleComponent } from './list-example/examples/selectable-items';
 import { ListExampleComponent } from './list-example/list-example.component';
 import { ListLoadOnDemandExampleComponent } from './list-load-on-demand-example/list-load-on-demand-example.component';
 import { ListNoShapeExampleComponent } from './list-no-shape-example/list-no-shape-example.component';
