+++ apps/cookbook/src/app/examples/list-example/examples/bold-text-on-row-selection.ts
