+++ apps/cookbook/src/app/examples/list-example/examples/colored-items.ts
