+++ apps/cookbook/src/app/examples/list-example/examples/header-and-footer.ts
