+++ apps/cookbook/src/app/examples/list-example/examples/items.ts
