+++ apps/cookbook/src/app/examples/list-example/examples/sections-and-colored-items.ts
