+++ apps/cookbook/src/app/examples/list-example/examples/selectable-items.ts
